committed

+++ dienquang_welcome/dashboard/save_thumbnail.php
@@ -1,12 +1,14 @@
 <?php
 
 include_once($_SERVER['DOCUMENT_ROOT'].'/database/web_content.php');
+include_once($_SERVER['DOCUMENT_ROOT'].'/dashboard/tools.php');
 
 $web_content = new WebContent();
 
+$rand_str = get_random_string(10);
 $embed_id = $_POST["id"];
 $target_dir = realpath(dirname(getcwd()))."\\dashboard\\upload\\";
-$target_file = $target_dir .'thumbnail_'. basename($_FILES["image_to_upload"]["name"]);
+$target_file = $target_dir .$rand_str. basename($_FILES["image_to_upload"]["name"]);
 $uploadOk = true;
 $imageFileType = strtolower(pathinfo($target_file, PATHINFO_EXTENSION));
 
@@ -34,15 +36,20 @@ try {
 if ($uploadOk){
   try {
     // Lấy content ảnh cũ ra và xóa
-    $old_data = mysqli_fetch_assoc($web_content->load_thumbnail($embed_id));
-    $dirname = explode("/",$old_data['thumbnail_link'])[1];
-    $filename = explode("/",$old_data['thumbnail_link'])[2];
-    unlink("./".$dirname."/".$filename);
+    $rs = $web_content->load_thumbnail($embed_id);
+    if ($rs != null){
+      $old_data = mysqli_fetch_assoc($rs);
+      if (!empty($old_data['thumbnail_link'])){
+        $dirname = explode("/",$old_data['thumbnail_link'])[1];
+        $filename = explode("/",$old_data['thumbnail_link'])[2];
+        unlink("./".$dirname."/".$filename);
+      }
+    }
   } catch (Exception $e) {
     echo "File ko tồn tại để xóa.";
   }
   // Lưu đường link
-  $data = '/upload/'.'thumbnail_'.$_FILES["image_to_upload"]['name'];
+  $data = '/upload/'.$rand_str.$_FILES["image_to_upload"]['name'];
   $web_content->save_thumbnail($embed_id, $data);
   // upload và đưa vào thư mục upload
   // tmp_name là cái file của host này đang chứa
